onyx: Fix missing torch tile for encrypted users
The torch tile is registered and appears fine when encryption isn't used
(thanks to several hacks inside the camera HAL), but the camera HAL hacks
are not holding up when encryption is used. The restart-media hack is still
needed right now to ensure the torch tile is registered successfully when
encryption is used; however, restarting media after boot is completed on
an unencrypted device causes the torch tile to go missing sometimes.

Introduce an init script to check if encryption is present in order to
restart media when it is needed, so that the torch tile appears all the time
for both encrypted and unencrypted users.